\name{mjmdscope-package}
\alias{mjmdscope-package}
\alias{mjmdscope}
\docType{package}
\title{\packageTitle{mjmdscope}}
\description{\packageDescription{mjmdscope}}
\section{Package Content}{\packageIndices{mjmdscope}}
\author{\packageAuthor{mjmdscope}}
\section{Maintainer}{\packageMaintainer{mjmdscope}}
\keyword{package}
